handle force/revoked/outdated close with splices
By default we assume that the final funding will make it to the
blockchain, which allows to use our closing procedure the same way as
without splices.

However, we still need to handle a corner case where a commitment tx for
one of the previous funding txs gets published and confirms. To deal
with that, we introduce a new `WatchAlternativeCommitTxConfirmed` which,
when triggers, causes all commitments but one to be pruned, allowing us
to go back to the normal case.