rework deactive/prune commitments
We make the following changes:
- only deactivate commitments when remote agrees
- only prune inactive commitments

Interestingly, this creates symmetry at two levels:
- between `updateLocalFundingStatus` and `updateRemoteFundingStatus`
- between `deactivateCommitments` and `pruneCommitments`

We also use the `WatchAlternativeCommitTxConfirmed` trick to handle
previous commitment been published for all states (it was only done in
`CLOSING`).

This creates two side-effects:
- in the rbf case, if a previous funding tx confirms but isn't locked by
remote, other competing funding txs won't be deactivated/pruned, and we
will first put a `WatchAlternativeCommitTxConfirmed` event (which will
fire immediately) even if we know it has already won the race.
- our simplistic tests on unknown spending txs don't work anymore,
because if we can't attach the transaction to a known commitment, we
assume that the related commitment was pruned and the event can safely be
ignored.