use typed actor for channel open replyTo
Not using `Status.Failure` anymore has impact on the api, because
akka's `ask` pattern translated them into a failed future, which in
turns resulted in an api 500 error. This behavior still remains in
corner cases (see test "open with bad arguments"), but otherwise we will
just return a successful future with a `OpenChannelResponse`. I think
this new behavior is fine, and in line with our long term goal of
getting rid of `ChannelException`.

In order to respect akka's guidelines we should make `replyTo` the first
argument of `OpenChannel`, but I didn't do it to minimize impact on
feature branches.

Note that there is a change in behavior between single and dual funding:
in the dual funding case we now reply with a success *before* attempting
to publish the signed funding transaction. The reason is that even if we
fail, our peer could succeed, and we really don't want a false negative
in scenarios where we are automatically opening channel.