rework response to rbf command
Main behavior change: we send the success response when the rbf flow is
successful, as opposed to when we initiate it. It requires adding the
command to `RbfStatus.InProgress` and allows a much more useful answer.

We use a typed `replyTo` actor, it's a first step towards reworking
other channel commands.

Note that we keep the `ChannelException` responses to minimize changes,
but in the future this should be a proper type hierarchy similar to
`OpenChannelResponse`.